AppServiceProvider and RouteServiceProvider moved on providers folder
